+++ packages/expo-updates/src/ExpoUpdatesModule.types.ts
@@ -48,6 +48,13 @@ export declare class ExpoUpdatesModule extends NativeModule<UpdatesEvents> {
   manifest?: Manifest;
   localAssets?: Record<string, string>;
 
+  initialContext: UpdatesNativeStateMachineContext & {
+    latestManifestString?: string;
+    downloadedManifestString?: string;
+    lastCheckForUpdateTimeString?: string;
+    rollbackString?: string;
+  };
+
   reload: () => Promise<void>;
   checkForUpdateAsync: () => Promise<
     | UpdateCheckResultRollBack
@@ -64,16 +71,5 @@ export declare class ExpoUpdatesModule extends NativeModule<UpdatesEvents> {
         ({ manifestString: string } | { manifest: Manifest }))
     | UpdateFetchResultFailure
     | UpdateFetchResultRollBackToEmbedded
-  >;
-  /**
-   * @hidden
-   */
-  getNativeStateMachineContextAsync: () => Promise<
-    UpdatesNativeStateMachineContext & {
-      latestManifestString?: string;
-      downloadedManifestString?: string;
-      lastCheckForUpdateTimeString?: string;
-      rollbackString?: string;
-    }
   >;
 }