+++ packages/expo-updates/src/UseUpdates.ts
@@ -1,13 +1,8 @@
 import { useEffect, useState } from 'react';
 
-import { getNativeStateMachineContextAsync } from './Updates';
-import { addUpdatesStateChangeListener } from './UpdatesEmitter';
+import { addUpdatesStateChangeListener, latestContext } from './UpdatesEmitter';
 import type { UseUpdatesReturnType } from './UseUpdates.types';
-import {
-  currentlyRunning,
-  defaultUseUpdatesState,
-  reduceUpdatesStateFromContext,
-} from './UseUpdatesUtils';
+import { currentlyRunning, updatesStateFromContext } from './UseUpdatesUtils';
 import type { UseUpdatesStateType } from './UseUpdatesUtils';
 
 /**
@@ -59,20 +54,12 @@ import type { UseUpdatesStateType } from './UseUpdatesUtils';
  * ```
  */
 export const useUpdates: () => UseUpdatesReturnType = () => {
-  const [updatesState, setUpdatesState] = useState<UseUpdatesStateType>(defaultUseUpdatesState);
+  const [updatesState, setUpdatesState] = useState<UseUpdatesStateType>(latestContext);
 
   // Change the state based on native state machine context changes
   useEffect(() => {
-    getNativeStateMachineContextAsync()
-      .then((context) => {
-        setUpdatesState((updatesState) => reduceUpdatesStateFromContext(updatesState, context));
-      })
-      .catch((error) => {
-        // Native call can fail (e.g. if in development mode), so catch the promise rejection and surface the error
-        setUpdatesState((updatesState) => ({ ...updatesState, initializationError: error }));
-      });
     const subscription = addUpdatesStateChangeListener((event) => {
-      setUpdatesState((updatesState) => reduceUpdatesStateFromContext(updatesState, event.context));
+      setUpdatesState(updatesStateFromContext(event.context));
     });
     return () => subscription.remove();
   }, []);